Support multiple links with same rel
This changes behavior of inherited links. Previously when a parent
serializer and a child serializer specified a link with the same rel,
then the child serializer overwrote this parent link.
Now that linke with same rels are combined into an array, that means
that both parent links and child links will be present in the array.